# Improve performance of str::to_lowercase and str::to_uppercase