@alexcrichton 2016-08-30
Thanks for the PR! Can you also be sure to add a few test cases for some of the more tricky corner cases this needs to handle as well?