@meh 2016-08-30
Where should I put them? I'm still not sure where the tests for stuff are supposed to go.

Do you have any specific examples you're concerned about? The only actual special case that was in the previous code is the final sigma one, which is basically copied as is.