@bluss 2016-08-31
There exists tests in src/libcollectionstest/str.rs already, check if they cover the special cases already (I think they do).