@meh 2016-09-01
The failing test is confusing, but my Unicode knowledge doesn't go that far.

If anyone could enlighten me, these characters `Ǆ ǅ ǆ Ǉ ǈ ǉ Ǌ ǋ ǌ` seem to have weird behavior with `is_uppercase` and `is_lowercase`, I could add an exception for it but I'm not sure there aren't any others with this behavior.