@frewsxcv 2016-09-01
From what I understand, unicode classifies characters as: uppercase, lowercase, and titlecase

* [Ǆ (uppercase)](http://www.fileformat.info/info/unicode/char/01c4/index.htm): is_uppercase=true, is_lowercase=false
* [ǅ (titlecase)](http://www.fileformat.info/info/unicode/char/01c5/index.htm): is_uppercase=false, is_lowercase=false
* [ǆ (lowercase)](http://www.fileformat.info/info/unicode/char/01c6/index.htm): is_uppercase=false, is_lowercase=true