@frewsxcv 2016-09-01
So in this test, which is currently failing:

```
assert_eq!("AÉǅaé ".to_lowercase(), "aéǆaé ");
```

It's expecting `ǅ` (titlecase) to be translated into its lowercase variant